Update HealthCare_Exchange_ML.ipynb
In LinearRegression() model, lm.score() score function used to find the goodness of the model. In statistics it is represented by "R-Squared" which is coefficient of determination, that shows the proportion of variance of dependent variable predictable from independent variables.
Have removed accuracy word from the comment section which was wrong interpretation.